refactor: add validate package
- Move Validator class to validate package
