+++ src/test/java/racingcar/ValidatorTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.validate.Validator;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
